backend/app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "google-cloud-aiplatform",
  "google-cloud",
  "google-cloud-logging",
  "pydantic_settings",
  "pandas",
  "python-dotenv",
  "db-dtypes",
  "google-adk",
  "moviepy>=2.2.1",
  "google-cloud-texttospeech>=2.32.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-mock>=3.15.1",
]
